Consolidate MIT operation model to one module
