+++ floris/core/turbine/operation_models.py
@@ -20,8 +20,6 @@ from floris.core.rotor_velocity import (
     rotor_velocity_air_density_correction,
     rotor_velocity_tilt_cosine_correction,
     rotor_velocity_yaw_cosine_correction,
-    mit_rotor_axial_induction,
-    mit_rotor_velocity_yaw_correction,
 )
 from floris.type_dec import (
     NDArrayFloat,
@@ -700,172 +698,3 @@ class PeakShavingTurbine():
         )
 
         return (1 - np.sqrt(1 - thrust_coefficient))/2
-
-
-class MITTurbine(BaseOperationModel):
-    """
-    Base class for turbine operation models. All turbine operation models must implement static
-    power(), thrust_coefficient(), and axial_induction() methods, which are called by power() and
-    thrust_coefficient() through the interface in the turbine.py module.
-
-    Args:
-        BaseClass (_type_): _description_
-
-    Raises:
-        NotImplementedError: _description_
-        NotImplementedError: _description_
-    """
-
-    @staticmethod
-    @abstractmethod
-    def power(
-        power_thrust_table: dict,
-        velocities: NDArrayFloat,
-        air_density: float,
-        yaw_angles: NDArrayFloat,
-        average_method: str = "cubic-mean",
-        cubature_weights: NDArrayFloat | None = None,
-        **kwargs,
-    ) -> None:
-
-
-        # Construct thrust coefficient interpolant
-        thrust_coefficient_interpolator = interp1d(
-            power_thrust_table["wind_speed"],
-            power_thrust_table["thrust_coefficient"],
-            fill_value=0.0001,
-            bounds_error=False,
-        )
-
-        # Compute the power-effective wind speed across the rotor
-        rotor_average_velocities = average_velocity(
-            velocities=velocities,
-            method=average_method,
-            cubature_weights=cubature_weights,
-        )
-
-        rotor_effective_velocities = rotor_velocity_air_density_correction(
-            velocities=rotor_average_velocities,
-            air_density=air_density,
-            ref_air_density=power_thrust_table["ref_air_density"]
-        )
-
-        thrust_coefficients = thrust_coefficient_interpolator(rotor_effective_velocities)
-
-
-        # mit_rotor_velocity_yaw_correction
-        axial_inductions = mit_rotor_axial_induction(thrust_coefficients, yaw_angles)
-
-
-        corrected_rotor_effective_velocities = mit_rotor_velocity_yaw_correction(
-           thrust_coefficients, yaw_angles, axial_inductions, rotor_effective_velocities
-        )
-
-        # Tilt correction?
-
-
-        # Construct power interpolant
-        power_interpolator = interp1d(
-            power_thrust_table["wind_speed"],
-            power_thrust_table["power"],
-            fill_value=0.0,
-            bounds_error=False,
-        )
-
-        # Compute power
-        power = power_interpolator(corrected_rotor_effective_velocities) * 1e3 # Convert to W
-
-        return power
-
-
-    @staticmethod
-    @abstractmethod
-    def thrust_coefficient(
-        power_thrust_table: dict,
-        velocities: NDArrayFloat,
-        air_density: float,
-        yaw_angles: NDArrayFloat,
-        average_method: str = "cubic-mean",
-        cubature_weights: NDArrayFloat | None = None,
-        **kwargs,
-    ) -> None:
-
-        # Construct thrust coefficient interpolant
-        thrust_coefficient_interpolator = interp1d(
-            power_thrust_table["wind_speed"],
-            power_thrust_table["thrust_coefficient"],
-            fill_value=0.0001,
-            bounds_error=False,
-        )
-
-        # Compute the power-effective wind speed across the rotor
-        rotor_average_velocities = average_velocity(
-            velocities=velocities,
-            method=average_method,
-            cubature_weights=cubature_weights,
-        )
-
-        rotor_effective_velocities = rotor_velocity_air_density_correction(
-            velocities=rotor_average_velocities,
-            air_density=air_density,
-            ref_air_density=power_thrust_table["ref_air_density"]
-        )
-
-        thrust_coefficients = thrust_coefficient_interpolator(rotor_effective_velocities)
-
-
-        # mit_rotor_velocity_yaw_correction
-        axial_inductions = mit_rotor_axial_induction(thrust_coefficients, yaw_angles)
-
-
-        corrected_rotor_effective_velocities = mit_rotor_velocity_yaw_correction(
-           thrust_coefficients, yaw_angles, axial_inductions, rotor_effective_velocities
-        )
-
-        # Tilt correction?
-
-        # Compute thrust coefficient
-        yawed_thrust_coefficients = thrust_coefficient_interpolator(corrected_rotor_effective_velocities) 
-
-        return yawed_thrust_coefficients
-
-    @staticmethod
-    @abstractmethod
-    def axial_induction(
-        power_thrust_table: dict,
-        velocities: NDArrayFloat,
-        air_density: float,
-        yaw_angles: NDArrayFloat,
-        average_method: str = "cubic-mean",
-        cubature_weights: NDArrayFloat | None = None,
-        **kwargs,
-    ):
-
-        # Construct thrust coefficient interpolant
-        thrust_coefficient_interpolator = interp1d(
-            power_thrust_table["wind_speed"],
-            power_thrust_table["thrust_coefficient"],
-            fill_value=0.0001,
-            bounds_error=False,
-        )
-
-        # Compute the power-effective wind speed across the rotor
-        rotor_average_velocities = average_velocity(
-            velocities=velocities,
-            method=average_method,
-            cubature_weights=cubature_weights,
-        )
-
-        rotor_effective_velocities = rotor_velocity_air_density_correction(
-            velocities=rotor_average_velocities,
-            air_density=air_density,
-            ref_air_density=power_thrust_table["ref_air_density"]
-        )
-
-        thrust_coefficients = thrust_coefficient_interpolator(rotor_effective_velocities)
-
-
-        # mit_rotor_velocity_yaw_correction
-        axial_inductions = mit_rotor_axial_induction(thrust_coefficients, yaw_angles)
-
-        return axial_inductions
